Bump minor and patch gem versions
