+++ app/controllers/manage_competencies/history_controller.rb
@@ -1,7 +1,7 @@
 module ManageCompetencies
   class HistoryController < ManageCompetencies::BaseController
     def show
-      @caseworker = user_scope.find(params[:id])
+      @caseworker = user_scope.find(params.expect(:id))
     end
   end
 end
