+++ app/controllers/manage_users/history_controller.rb
@@ -1,7 +1,7 @@
 module ManageUsers
   class HistoryController < ManageUsers::BaseController
     def show
-      @user = User.find(params[:id])
+      @user = User.find(params.expect(:id))
     end
   end
 end
