+++ app/lib/govuk_design_system_form_builder/elements/inputs/search_date.rb
@@ -1,0 +1,20 @@
+module GOVUKDesignSystemFormBuilder
+  module Elements
+    module Inputs
+      class SearchDate < Base
+        include Traits::Input
+        include Traits::Error
+        include Traits::Hint
+        include Traits::Label
+        include Traits::Supplemental
+        include Traits::HTMLAttributes
+
+        private
+
+        def builder_method
+          :date_field
+        end
+      end
+    end
+  end
+end
