+++ spec/system/casework/reviewing/mark_application_as_ready_spec.rb
@@ -112,7 +112,7 @@ RSpec.describe 'Marking an application as ready for assessment' do
     end
 
     it 'returns a 409 error status' do
-      expect(page).to have_http_status :unprocessable_entity
+      expect(page).to have_http_status :unprocessable_content
     end
 
     it 'uses the system user layout with navigation' do
