+++ spec/system/casework/reviewing/send_back_to_provider_spec.rb
@@ -117,7 +117,7 @@ RSpec.describe 'Send an application back to the provider' do
         end
 
         it 'returns a 409 error status' do
-          expect(page).to have_http_status :unprocessable_entity
+          expect(page).to have_http_status :unprocessable_content
         end
 
         it 'uses the system user layout with navigation' do
